Finish Elevenlabs modularization
